add camera

+++ src/com/example/BGL/object/Bobject.java
@@ -55,10 +55,10 @@ public class Bobject {
         pressed = false;
     }
     //When moving your finger over the object
-    public void touchMove(int x, int y) {
+    public void touchMove(float x, float y) {
     }
     // Triggered when releasing the finger and it's not over the object anymore
-    public void touchUpMove(int x, int y){
+    public void touchUpMove(float x, float y){
         pressed = false;
     }
